+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Collection loader
 - Ethereum bridge  tab
 - Metamask form
-- Web3 provider module 
+- Web3 provider module
+- Link to open this site of the metamask mobile application
 
 #### Changed
 - Version api packages